+++ project_name/settings/dev.py
@@ -13,7 +13,7 @@ from .base import *  # noqa
 
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = '{{ secret_key }}'
+SECRET_KEY = os.environ.get('SECRET_KEY', 'not-safe-for-prod')
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True