+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Version 3.1
 In development.
 
 * Updated psycopg2-binary to version 2.8.4.
+* Moved flake8 config from ``setup.cfg`` to ``tox.ini``.
 
 ----