+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ In development.
 
 * Updated psycopg2-binary to version 2.8.4.
 * Moved flake8 config from ``setup.cfg`` to ``tox.ini``.
+* Renamed ̀``settings/common.py`` as ``settings/base.py``.
 
 ----